% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaSim.R
\name{arimaSim}
\alias{arimaSim}
\title{ARIMA process realization}
\usage{
arimaSim(n = 64, ar = NULL, ma = NULL, int = 0, ...)
}
\arguments{
\item{n}{number of output samples}

\item{ar}{vector of ar coefficients; defaults to NULL (ignored)}

\item{ma}{vector of ma coefficients; defaults to NULL (ignored)}

\item{int}{integration order, a non-negative integer; defaults to 0 (no integration)}

\item{...}{other arguments passed to arima.sim.  sd is the most useful}
}
\value{
ts object with n + int samples.
}
\description{
Produce realization of an ARIMA process using \code{\link{arima.sim}}.
}
\examples{
a <- arimaSim( n = 100, ar = 0.9, int = 1 )
plot( a )
b <- arimaSim( n = 100, ma = c( 0.5, 0.1 ), ar = c( -0.2 ) )
plot( b )

}
