% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps.R
\name{arps}
\alias{arps}
\title{Compute theorectical AR(p) power spectrum}
\usage{
arps(p = 1, a = c(0.9), freq = 100)
}
\arguments{
\item{p}{process order (defaults to 1)}

\item{a}{single coefficient (p = 1) or vector of coefficients, a1, a2, ..., ap.
No check is made to insure stationarity.}

\item{freq}{either a scalar specifying how many frequencies between 0 and 0.5 to return
or a vector of frequencies.}
}
\value{
AR(p) power spectral density
}
\description{
Compute theorectical AR(p) power spectrum.
}
\examples{
f <- seq( 0, 0.5, 0.01 )
s <- arps( p = 1, a = 0.5, freq = f )
plot( f, s, type = "l" )
}
