% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fht.R
\name{binary2decimal}
\alias{binary2decimal}
\title{binary2decimal}
\usage{
binary2decimal(x)
}
\arguments{
\item{x}{binary representation as vector of digits (0,1)}
}
\value{
integer value
}
\description{
Convert binary digits vector to non-negative integer value
}
\keyword{internal}
