% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cepstrum.R
\name{cepstrum}
\alias{cepstrum}
\title{Cepstrum of univariate real timeseries}
\usage{
cepstrum(s, taper = c("hamming", "akaike", "gausswin"), poly = 5)
}
\arguments{
\item{s}{real-valued, univariate timeseries}

\item{taper}{taper applied to s prior to fft; one of "hamming" (default), "akaike" or "gausswin"}

\item{poly}{polynomial order of trend removed from the log of power spectral density; defaults to 5}
}
\value{
data.frame with quefrency and cepstrum
}
\description{
Compute cepstrum of a univariate, real-valued timeseries
}
\examples{
t <- 1:500
f <- seq( 0.05, 0.25, 0.05 )
s <- sapply( t, function( x, y ) sum( sin( 2 * pi * x * y ) ), y = f )
s <- s + rnorm( length( t ), sd = 0.5 )
c <- cepstrum( s, taper = "akaike" )
ggplot( c, aes( x = quefrency, y = cepstrum ) ) + geom_line() +
labs( x = "Quefrency", y = "Cepstrum" )
# harmonics spaced at 0.05 produce peak at quefrency of 20 (1 / 20 = 0.05 )

}
