% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosfilt.R
\name{cosfilt}
\alias{cosfilt}
\title{Cosine-squared timeseries filtering}
\usage{
cosfilt(s, fc = c(0, 0.2, 0.3, 0.5), fw = c(1, 1, 0, 0))
}
\arguments{
\item{s}{univariate or multivariate (as matrix columns) real-valued timeseries.}

\item{fc}{vector of "corner" frequencies specifying endpoints of frequency bands.
Regardless of sample rate, these are specified in terms of the Nyquist band
(0,0.5).}

\item{fw}{amplitudes of filter at corresponding corner frequencies.  Lengths of fc and fw must match.
Must be 0 (reject) or 1 (pass).}
}
\value{
filtered timeseries
}
\description{
Filter a regularly sampled univariate or multivariate timeseries using cosine-squared,
symmetric filter bank.  Filtering is performed in the frequency domain.
}
\details{
By specifying fc and fw, one can create many filters: low-pass, high-pass, band-pass
and band-reject as well as their combinations.  Filter response will be 0 or 1 except across
intervals where fw changes.  The lengths of these transitions affect the length of the effective filter impulse response.  
Short transitions produce long filters.  Filtering is performed in the frequency domain and is circular. 
Input timeseries is padded to next factor of 2 prior to filtering. If the length is near a factor of 2,
you may want to pad s prior to use of cosfilt, especially if filter transition intervals are short.
}
\examples{
s <- 0.5 * sin( 2 * pi * 1:200 / 20 ) + rnorm( 200 )
fs <- cosfilt( s, fc = c( 0, 0.1, 0.2, 0.5 ), fw = c( 1, 1, 0, 0 ) )
plot( s, type = "l" )
lines( fs, col = "red" )

}
