% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftresample.R
\name{costap}
\alias{costap}
\title{Cosine-squared taper}
\usage{
costap(n, pct = 0.05)
}
\arguments{
\item{n}{length of taper}

\item{pct}{percent of n used to roll on and off the taper on ends.}
}
\value{
taper vector.
}
\description{
Compute cosine-squared taper
}
\examples{
t <- costap( 100, pct = 0.2 )
plot( t, xlab = "Time", ylab = "Taper", type = "l" )
}
