% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fht.R
\name{decimal2binary}
\alias{decimal2binary}
\title{decimal2binary}
\usage{
decimal2binary(x, length)
}
\arguments{
\item{x}{non-negative integer to convert}

\item{length}{length of binary sequence to output}
}
\value{
binary representation as vector of digits (0,1)
}
\description{
Convert non-negative integer to binary
}
\keyword{internal}
