% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Detrend a time series with a low-order polynomial.}
\usage{
detrend(x, n = 1)
}
\arguments{
\item{x}{time series or vector of data values}

\item{n}{polynomial order}
}
\value{
ts object with low-order LS polynomial removed.
}
\description{
Remove low-order polynomial "trend" from a time series.
}
\examples{
n <- 100
x <- arimaSim( n, ar = 0.9, int = 1 ) + 1:n / 20
plot( x )
d <- detrend( x, 1 )
plot( d )

}
