% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttools.R
\name{df2ts}
\alias{df2ts}
\title{Convert data.frame object to a ts}
\usage{
df2ts(df, time)
}
\arguments{
\item{df}{data.frame with regularly sampled time series coefficients as columns and time as an additional column.}

\item{time}{character vector name of time column or its numeric index.}
}
\value{
ts object with converted data.frame object.
}
\description{
Convert data.frame object to a ts
}
\examples{
s.df <- data.frame( time = 1:100, s = sin( 2 * pi * 1:100 / 20 ) )
s.ts <- df2ts( s.df, "time" )

}
