% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{envelope}
\alias{envelope}
\title{Envelope using analytic signal}
\usage{
envelope(s)
}
\arguments{
\item{s}{vector of timeseries samples or ts object.
NA values are not allowed.}
}
\value{
Real valued (non-negative) envelope of x.
}
\description{
Compute envelope (instantaneous amplitude) of a time series using hilbert transform
}
\examples{
s <- sin( 2 * pi * 1:256 * 0.05 )
S <- envelope( s )
plot( S, xlab = "Time", ylab = NULL, type = "l", ylim = c( -1.1, 1.1 ) )
lines( s, col = "red" )
}
