% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palinsol_helper.R
\name{esin}
\alias{esin}
\title{esin}
\usage{
esin(orb)
}
\arguments{
\item{orb}{orbital calculation output from one of \code{\link{la04}}, \code{\link{ber78}} or \code{\link{ber90}}}
}
\description{
Compute esin term used in precession index computation.  There should be no need to call this separate from \code{\link{precession}}
}
\keyword{internal}
