% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutive.R
\name{evolFFT}
\alias{evolFFT}
\title{Time-Frequency Analysis}
\usage{
evolFFT(s, deltat = 1, nfft = 1024, nwin, nstep, fl = 0, fh = 0.5/deltat)
}
\arguments{
\item{s}{vector of timeseries samples or ts object.  If the former, deltat should be specified,
NA values are not allowed.}

\item{deltat}{sampling interval of time series.  Ignored if s is a ts object}

\item{nfft}{length of padding window input to fft; controls frequency resolution}

\item{nwin}{number of samples in fft window}

\item{nstep}{number of samples to shift right between spectral estimates; controls time resolution}

\item{fl}{lowest frequency to output}

\item{fh}{highest frequence to output}
}
\value{
List containing the following items:
 \item{"signal"}{Original input signal}
 \item{"deltat"}{Sampling interval of time series}
 \item{"numfreqs"}{Number of frequencies retained in analysis}
 \item{"freqs"}{Vector of frequencies retained in analysis}
 \item{"wpars"}{List of analysis parameters: fl, fh and nwin}
}
\description{
Compute evolutive spectrum using direct FFT method
}
\examples{
require( RColorBrewer )
# create chirp signal
n <- 1024
t <- 1:n
a <- 4 * n
x <- cos( 2 * pi * t^2 / a )
# perform time-frequency analysis of chirp
evol <- evolFFT( x, deltat = 1, nwin = 200, nstep = 5 )
# plot result
plotFTSpec( evol )
}
