% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdfilter.R
\name{fdfilter}
\alias{fdfilter}
\title{Finite difference stencil filter}
\usage{
fdfilter(n = 3, d = 1)
}
\arguments{
\item{n}{total length of filter.  3 <= n < 19 and odd.}

\item{d}{order of the derivative.  Must be an integer and satisfy
1 <= d < \code{n}}
}
\value{
Derivative filter coefficients.  Use \code{\link[stats]{filter}}
two-sided filter.
}
\description{
Compute a finite-difference "stencil" for numerical differentiation
}
\examples{
f <- fdfilter( 5, 1 )
time <- 1:100
period <- 20
phase <- 2 * pi * time / period
s <- sin( phase )
Ds <- stats::filter( s, f )
plot( Ds, type = "l" )
lines( 2 * pi / period * cos( phase ), col = "red" )

}
