% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftresample.R
\name{fftdecimate}
\alias{fftdecimate}
\title{Downsample timeseries}
\usage{
fftdecimate(s, nfac = 2, pct = 0.025, deltat = 1)
}
\arguments{
\item{s}{timeseries (as a ts object or vector of samples)}

\item{nfac}{downsampling factor (integer greater than 1)}

\item{pct}{percent length of timeseries used for cosine-squared taper of ends.
This reduces edge effects, but is lossy.}

\item{deltat}{sampling interval; used only if s is not a ts object}
}
\value{
resampled timeseries.
}
\description{
Downsample a timeseries using FFT
}
\examples{
s <- ts( sin( 2 * pi * 1:1000 / 200 ), deltat = 0.1 )
ds <- fftdecimate( s, nfac = 10 )
plot( s, type = "l" )
lines( ds, col = "red" )
}
