% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttools.R
\name{fftnneg}
\alias{fftnneg}
\title{Non-negative frequencies of FFT}
\usage{
fftnneg(s)
}
\arguments{
\item{s}{vector of time series samples}
}
\value{
complex valued fft of s at non-negative frequencies, ordered DC to Nyquist
}
\description{
Compute non-negative frequency coefficients of fft
}
\examples{
s <- sin( 2 * pi * ( 1:256 ) * 0.1 )
S <- fftnneg( s )
f <- seq( 0, to = 0.5, length.out = length( S ) )
plot( f, Mod( S ), type = "l", xlab = "Frequency", ylab = "Mod(S)" )

}
