% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftresample.R
\name{fftresample}
\alias{fftresample}
\title{Upsample timeseries}
\usage{
fftresample(s, nfac = 2, pct = 0.025, deltat = 1)
}
\arguments{
\item{s}{timeseries (as a ts object or vector of samples)}

\item{nfac}{upsampling factor (integer greater than 1)}

\item{pct}{percent length of timeseries used for cosine-squared taper of ends.
This reduces edge effects, but is lossy.}

\item{deltat}{sampling interval; used only if s is not a ts object}
}
\value{
resampled timeseries.
}
\description{
Upsample a timeseries using FFT
}
\examples{
s <- ts( sin( 2 * pi * 1:100 / 20 ), deltat = 1 )
us <- fftresample( s, nfac = 10 )
plot( s, type = "l" )
lines( us, col = "red" )
}
