% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fht.R
\name{fht}
\alias{fht}
\title{Fast Hadamard Transform}
\usage{
fht(x)
}
\arguments{
\item{x}{vector of univariate sequence values; zero padded to next multiple of 2}
}
\value{
vector of Hadamard transform coefficients
}
\description{
"Natural" order Hadamard transform.
}
\examples{
a <- arimaSim( n = 124, ar = 0.9, int = 1 )
H <- fht( a )
plot( H, type = "l", xlab = "Index", ylab = "Coefficient" )

}
