% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz.R
\name{fviz}
\alias{fviz}
\title{Plot amplitude and phase of linear filter}
\usage{
fviz(
  h,
  type = c("centered", "causal", "symmetric", "arma"),
  phase = TRUE,
  hide = FALSE,
  twoway = FALSE
)
}
\arguments{
\item{h}{filter coefficients; see details below}

\item{type}{one of "centered", "causal", "symmetric" or "arma."
The first three types assume \code{h} is a vector of filter coefficients.
If type is "centered" these are centered around zero.  If the filter length is even,
the filter extends further to positive time.  Type "causal" assumes coefficients start
with time 0.  Type "symmetric" assumes \code{h} provides the time >= 0 coefficients.  
These are mirrored for negative times.
Type "arma" is for filters created in the \code{\link{signal}} package.}

\item{phase}{include filter phase spectrum facet if true}

\item{hide}{if true, function returns a list with with filter z transform (see below)
and ggplot object}

\item{twoway}{if true and filter type is ARMA, response is computed for two-way filter}
}
\value{
List with frequency "f" and z representation of filter "H"
}
\description{
Plot amplitude and phase of a linear filter
}
\examples{
f <- fdfilter( 5, 1 )
H <- fviz( f )

}
