% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fht.R
\name{fwt}
\alias{fwt}
\title{Fast Walsh Transform}
\usage{
fwt(x)
}
\arguments{
\item{x}{vector of univariate sequence values; zero padded to next multiple of 2}
}
\value{
vector of Walsh transform coefficients
}
\description{
Fast Walsh transform (aka "Sequency" order Hadamard)
}
\examples{
a <- arimaSim( n = 124, ar = 0.9, int = 1 )
H <- fwt( a )
plot( H, type = "l", xlab = "Index", ylab = "Coefficient" )

}
