% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.acf.R
\name{gplot.ccf}
\alias{gplot.ccf}
\title{Plot cross-correlation function}
\usage{
gplot.ccf(
  x,
  y,
  lag.max = NULL,
  type = c("correlation", "covariance"),
  demean = TRUE,
  ci = 0.95
)
}
\arguments{
\item{x}{a timeseries}

\item{y}{a timeseries}

\item{lag.max}{maximum time lag to display}

\item{type}{one of "correlation" (default) or "covariance."
The latter is not implemented.}

\item{demean}{remove timeseries mean prior to computation of ACF}

\item{ci}{confidence interval to display}
}
\value{
a ggplot2 object
}
\description{
Attractive \link{ggplot2} plot of \code{\link{ccf}} output
}
\examples{
x <- rnorm( 1000 )
plot.acf( x, lag.max = 20 )
}
