% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.filter.R
\name{gplot.filter}
\alias{gplot.filter}
\title{Plot short convolution filter}
\usage{
gplot.filter(h, type = c("centered", "causal", "symmetric"))
}
\arguments{
\item{h}{a filter}

\item{type}{one of "centered" (default), "causal" or "symmetric".
If "centered" filter is shifted left to center on origin.  If even length, it
will extend further to the right of 0.  If symmetric, h is assumed to provide only
the time >= 0 coefficients.  Negative time will be obtained by mirroring.}
}
\value{
a ggplot2 object
}
\description{
Attractive \link{ggplot2} plot of short convolution filter coefficients
}
\examples{
h <- fdfilter( 5 )
gplot.filter( h )
}
