% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_plots.R
\name{gplot.mtm.coh}
\alias{gplot.mtm.coh}
\title{Plot coherence results from \code{\link{mtm.coh}}}
\usage{
gplot.mtm.coh(mcoh, uci = 0.95, period = F, trans = c("identity", "atanh"))
}
\arguments{
\item{mcoh}{\code{\link{mtm.coh}} output}

\item{uci}{upper confidence interval to highlight in plot}

\item{period}{display MSC against period, rather than frequency, if TRUE}

\item{trans}{apply atanh (aka Fisher) transform to the MSC; defaults to no transform}
}
\value{
a ggplot2 object
}
\description{
Attractive \link{ggplot2} plot of \code{\link{spec.mtm}} F-test output
}
\examples{
s <- ts( sin( 2 * pi * 1:512 / 32 ) + rnorm( 512 ), deltat = 1 )
mspec <- spec.mtm( s, Ftest = TRUE, plot = FALSE )
mtm.ftest( mspec )

}
