% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsplot.R
\name{gplot.ts}
\alias{gplot.ts}
\title{Plot timeseries}
\usage{
gplot.ts(s)
}
\arguments{
\item{s}{a timeseries}
}
\value{
a ggplot2 object
}
\description{
Attractive \link{ggplot2} plot of an univariate timeseries
}
\examples{
Timeseries <- ts( rnorm( 100 ), deltat = 1 )
gtsplot( Timeseries )

}
