% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fht.R
\name{gray2binary}
\alias{gray2binary}
\title{gray2binary}
\usage{
gray2binary(x)
}
\arguments{
\item{x}{gray code as vector of digits (0,1)}
}
\value{
binary representation as vector of digits (0,1)
}
\description{
Convert gray code to binary
}
\keyword{internal}
