% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{hilbert}
\alias{hilbert}
\title{Hilbert transform}
\usage{
hilbert(s)
}
\arguments{
\item{s}{vector of timeseries samples or ts object.
NA values are not allowed.}
}
\value{
Complex valued hilbert transform of x.
}
\description{
Compute hilbert transform using FFT
}
\examples{
s <- sin( 2 * pi * 1:256 * 0.05 )
H <- hilbert( s )
plot( Re( H ), xlab = "Time", ylab = NULL, type = "l", ylim = c( -1.1, 1.1 ) )
lines( Im( H ), col = "red" )
}
