% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos}
\alias{lanczos}
\title{Create a Lanczos filter}
\usage{
lanczos(n = 7, type = c("lp", "hp", "bp", "br"), low = 0, high = 0.5)
}
\arguments{
\item{n}{length of filter; must be odd}

\item{type}{one of "lp" (low pass), "hp" (high pass), "bp" (band pass) or "br" (band reject)}

\item{low}{low frequency transition of "lp", "bp" and "br" filters (Nyquist = 0.5)}

\item{high}{high frequency transition of "hp", "bp" and "br" filters (Nyquist = 0.5)}
}
\value{
non-negative time filter coefficients (filter is symmetric about 0).
}
\description{
Compute coefficients of a Lanczos filter
}
\examples{
# low-pass filter
lp <- lanczos( 7, "lp", low = 0.1 )

# high-pass filter
hp <- lanczos( 7, "hp", high = 0.4 )

# band-pass filter
bp <- lanczos( 7, "bp", low = 0.1, high = 0.4 )

# band-reject filter
br <- lanczos( 7, "br", low = 0.1, high = 0.4 )

plot( lp, xlab = "Index", ylab = "Filter Coefficients", type = "l", ylim = 1.5 * range( lp, hp, bp, br ) )
lines( hp, col = "red" )
lines( bp, col = "blue" )
lines( br, col = "green" )
}
