% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_br}
\alias{lanczos_br}
\title{Create a band-reject Lanczos filter}
\usage{
lanczos_br(n, low, high)
}
\arguments{
\item{n}{length of filter; must be odd}

\item{low}{transition frequency of the filter (Nyquist = 0.5)}

\item{high}{transition frequency of the filter (Nyquist = 0.5)}
}
\value{
non-negative time filter coefficients (filter is symmetric about 0). 
Frequencies inside band (low, high) are attenuated.
}
\description{
Compute coefficients of a band-reject Lanczos filter
}
\examples{
f <- lanczos_br( 7, 0.1, 0.4 )
print( f )

}
