% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_hp}
\alias{lanczos_hp}
\title{Create a high-pass Lanczos filter}
\usage{
lanczos_hp(n, high)
}
\arguments{
\item{n}{length of filter; must be odd}

\item{high}{cutoff frequency of the filter (Nyquist = 0.5)}
}
\value{
non-negative time filter coefficients (filter is symmetric about 0).
}
\description{
Compute coefficients of a high-pass Lanczos filter
}
\examples{
f <- lanczos_hp( 7, 0.4 )
print( f )

}
