% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttools.R
\name{leakage}
\alias{leakage}
\title{Spectral leakage}
\usage{
leakage(w)
}
\arguments{
\item{w}{windowing function coefficients}
}
\value{
log(modulus) of the padded fft
}
\description{
Evaluate side-lobe leakage for a given windowing function
}
\examples{
s <- rep( 1, 32 )
S <- leakage( s )
f <- seq( 0, to = 0.5, length.out = length( S ) )
plot( f, S, type = "l", xlab = "Frequency", ylab = "log(Mod(S))" )

}
