% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps.R
\name{maps}
\alias{maps}
\title{Compute theorectical MA(q) power spectrum.}
\usage{
maps(q = 1, b = c(0.9), freq = 100)
}
\arguments{
\item{b}{single coefficient (a = 1) or vector of coefficients, b1, b2, ..., bq.
No check is made to insure stationarity.}

\item{freq}{either a scalar specifying how many frequencies between 0 and 0.5 to return
or a vector of frequencies.}

\item{a}{process order (defaults to 1)}
}
\value{
MA(q) power spectral density
}
\description{
Compute theorectical MA(q) power spectrum.
}
\examples{
f <- seq( 0, 0.5, 0.01 )
s <- maps( p = 1, a = 0.5, freq = f )
plot( f, s, type = "l" )
}
