% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morlet.R
\name{morlet}
\alias{morlet}
\title{Morlet wavelet}
\usage{
morlet(n = 500, center = 250, scale = 10, w0 = 2 * pi)
}
\arguments{
\item{n}{number of output samples}

\item{center}{where to center the wavelet in n-sample sequence}

\item{scale}{dilation factor}

\item{w0}{sinusoid radial frequency (defaults to 2 * pi )}
}
\value{
vector of wavelet coefficients
}
\description{
Compute real-value morlet wavelet
}
\examples{
w <- morlet( 500, 250, 10 )
plot( 1:500, w, type = "l" )
}
