% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtoutlier.R
\name{mtoutlier}
\alias{mtoutlier}
\title{Martin-Thomson outlier removal}
\usage{
mtoutlier(y, ar = c(1, 10), tol = 3, set0 = TRUE, ...)
}
\arguments{
\item{y}{timeseries; no NA allowed}

\item{ar}{c(min, max) AR order in modeling; defaults to c(1, 10)}

\item{tol}{outlier threshold value; defaults to 3 (Pearson's Rule)}

\item{set0}{replace outlier with 0 (TRUE, default) or scaled tol (FALSE)}

\item{...}{other arguments passed to \code{\link{arima}}}
}
\value{
cleaned timeseries
}
\description{
Martin-Thomson filter-cleaning of a (presumed) AR timeseries
}
\examples{
y <- arimaSim( 200, ar = c( 0.5, -0.1 ) )
# Add a few outliers
o <- floor( runif( 5, min = 5, max = 195 ) )
y[o] <- runif( 5, min = 4, max = 8 )
yc <- mtoutlier( y )
plot( y )
lines( yc, col = "red" )

}
