% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcotes.R
\name{newcotes}
\alias{newcotes}
\title{Newton-Cotes integration filters}
\usage{
newcotes(n = 2)
}
\value{
filter matrix
}
\description{
Compute  matrix of stable Newton-Cotes integration filters up to order 7
}
