% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_tools.R
\name{pcaPPC}
\alias{pcaPPC}
\title{Principal component plot}
\usage{
pcaPPC(p, which = 1:length(p$sdev), scale = FALSE)
}
\arguments{
\item{p}{"prcomp" list from \code{\link{prcomp}}}

\item{which}{a vector with index vector of which PCs to show}

\item{scale}{if TRUE, scale PCs by the reciprocal of standard deviation.  If TRUE
PCs share a common scale.  If FALSE (default) scale depends on variance reduction.}
}
\description{
Plot (optionally) scaled principal components produced by \code{\link{prcomp}}
}
\examples{
# Not the best example--abscissa axes are very crowded.
p <- prcomp(USArrests, scale = TRUE)
pcaPPC( p, scale = TRUE )

}
