% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_tools.R
\name{pcaVar}
\alias{pcaVar}
\title{Cumulative variance reduction}
\usage{
pcaVar(p, plot = TRUE, n = length(p$sdev))
}
\arguments{
\item{p}{"prcomp" list from \code{\link{prcomp}}}

\item{plot}{produce a ggplot2 plot}

\item{n}{starting with PC #1, the number of PCs that should be included in results, defaults to all}
}
\value{
data.frame with PC index, the associated eigenvalue, variance (squared eigenvalue) and cumulative variance.
}
\description{
Compute and (optionally) plot cumulative variance reduction of PCs returned from \code{\link{prcomp}}
}
\examples{
p <- prcomp(USArrests, scale = TRUE)
pcaVar( p )

}
