% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFTSpec.R
\name{plotFTSpec}
\alias{plotFTSpec}
\title{Plot fourier/time spectral analysis}
\usage{
plotFTSpec(
  evol,
  Log = F,
  fscale = c("Frequency", "Octave", "Period"),
  pal = colorRampPalette(RColorBrewer::brewer.pal(11, "Spectral"))(256),
  xlab = "Time",
  ylab = "Frequency",
  dynRange = 0
)
}
\arguments{
\item{evol}{return value of \code{\link{evolFFT}} or \code{\link{evolMTM}}}

\item{Log}{plot log10 of power spectrum, defaults to FALSE}

\item{fscale}{one of "Frequency" (default), "Period" or "Octave" The latter two display period on a log10 scale and frequency on a log2 scale respectively. plot log2 of frequency, defaults to FALSE}

\item{pal}{color palette, defaults to a dark spectrum}

\item{xlab}{label for time axis, defaults to "Time"}

\item{ylab}{label for frequency/period axis, defaults to "Frequency" or "Period" as appropriate}

\item{dynRange}{orders of magnitude beneath peak to include in color scale.  Defaults to full range.}
}
\description{
Plot output of \code{\link{evolFFT}} and \code{\link{evolMTM}} in \code{\link{TSAUMN}}
Plot output of \code{\link{evolFFT}} and \code{\link{evolMTM}} in \code{\link{TSAUMN}}
}
\examples{
require( RColorBrewer )
# create chirp signal
n <- 1024
t <- 1:n
a <- 4 * n
x <- cos( 2 * pi * t^2 / a )
# perform time-frequency analysis of chirp
evol <- evolFFT( x, deltat = 1, nwin = 200, nstep = 5 )
# plot result
plotFTSpec( evol )
}
