% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotl.R
\name{plotl}
\alias{plotl}
\title{Shorthand line plotting}
\usage{
plotl(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{plot}}, including \code{\link{graphical parameters}} (see \code{\link{par}} ).}
}
\description{
Equivalent to plot( ..., type = "l" )
}
\examples{
x <- 1:50
y <- sin( 2 * pi * x / 10 ) + rnorm( 50 )
plotl( x, y, col = "blue" )
}
