% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTS_smoother.R
\name{RTSsmooth}
\alias{RTSsmooth}
\title{Rauch–Tung–Striebel Kalman smoothing}
\usage{
RTSsmooth(z, q = 0.5, r = 0.5)
}
\arguments{
\item{z}{univariate timeseries (ts) or single vector of values}

\item{q}{estimated process noise variance (V in StructTS$model)}

\item{r}{estimated measurement noise variance (h in StructTS$model)}
}
\value{
smoothed timeseries.
}
\description{
RTS kalman filter (smooth) a univariate timeseries with F and H = 1.  Simple function
to demonstrate kalman smoothing.
}
\examples{
z <- ts( sin( 2 * pi * 1:200 / 20 ) + rnorm( 200 ), deltat = 1 )
s <- RTSsmooth( z, q = 0.1, r = 1 )
plot( z, type = "l" )
lines( s, col = "red" )
}
