% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fht.R
\name{sequency}
\alias{sequency}
\title{Reorder vector into sequency order}
\usage{
sequency(n)
}
\arguments{
\item{n}{number of samples in sequence}
}
\value{
permutation index
}
\description{
Reorder vector into sequency order for Walsh transform using Fast Hadamard Transform \code{link{fht}}.
}
\keyword{internal}
