% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.bt.R
\name{spec.bt}
\alias{spec.bt}
\title{Blackman-Tukey spectral density estimation}
\usage{
spec.bt(
  s = NA,
  q = 20,
  taper = TRUE,
  n = 2048,
  deltat = 1,
  demean = TRUE,
  detrend = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{s}{univariate real-valued timeseries.}

\item{q}{maximum lag retained in the auto-correlation function.}

\item{taper}{apply Akaike ACF taper if true (default).}

\item{n}{length of zero-padded ACF function.}

\item{deltat}{sample interval of timeseries if s passed as a vector.}

\item{demean}{logical.  If true, timeseries mean is removed.}

\item{detrend}{logical.  If true, linear trend is removed from timeseries.}

\item{plot}{logical.  If true, spectrum is plotted using gplot.mtm}

\item{...}{Additional parameters passed to gplot.mtm}
}
\value{
list with items:
  \item{freq}{Vector of sampled frequencies}
  \item{spec}{Vector of power spectral estimates}
  \item{series}{Name of the input time series}
  \item{Method}{"Blackman Tukey"}
}
\description{
Perform Blackman-Tukey power-spectral density estimate of a timeseries.
}
\details{
Blackman-Tukey is a parametric power-spectral density estimator.
It assumes an MA(q) process with known q. In practice, increasing q
trades off greater spectral resolution with greater estimator variance. In all cases,
q should be much shorter than the timeseries itself.
}
\examples{
s <- arimaSim( 1028, ma = 0.5 )
spec <- spec.bt( s, q = 5 )

}
