% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttools.R
\name{ts2df}
\alias{ts2df}
\title{Convert ts object to a data.frame}
\usage{
ts2df(s)
}
\arguments{
\item{s}{time series object}
}
\value{
data.frame with time as column 1, time series as remaining columns.
}
\description{
Convert ts object to a data.frame
}
\examples{
s <- ts( data = sin( 2 * pi * (1:100) * 0.1 ), start = 1, end = 100, deltat = 1 )
S.df <- ts2df( s )
head( S.df )

}
