% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttools.R
\name{welch}
\alias{welch}
\title{Welch window}
\usage{
welch(n)
}
\arguments{
\item{n}{length of the filter; number of coefficients to generate.}
}
\value{
Filter coefficients
}
\description{
Welch's windowing function
}
\examples{
n <- 100
plot( welch( n ), type = "l" )

}
