% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whfilter.R
\name{whfilt}
\alias{whfilt}
\title{Apply derived FIR Causal Wiener-Hopf Filter}
\usage{
whfilt(x, s, h)
}
\arguments{
\item{x}{a ts, vector or matrix of values (columns as time series).}

\item{s}{as ts or vector of values}

\item{h}{list returned from \code{whfir}.}
}
\value{
Wiener-Hopf prediction of time series s.
}
\description{
Apply the optimal (MMSE) Wiener-Hopf causal FIR filter to a known multivariate
input ts x.  Uses filter output from \code{whfir}.
}
\examples{
x <- rnorm( 504 )
h <- seq( 1, 0, length.out = 5 )
s <- stats::filter( x, h, sides = 1 )
x <- x[-(1:4)] # Remove filter roll-on
s <- s[-(1:4)]
wh <- whfir( x, s, n = c(1, 10 ) )
spred <- whfilt( x, s, wh )
plot( s, type = "l" )
lines( spred, col = "red" )

}
