% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbispec_mtm.R
\name{xbispec.mtm}
\alias{xbispec.mtm}
\title{Cross-bispectrum}
\usage{
xbispec.mtm(x, y, deltat = 1, nw = 4, k = 7, nPoly = 3)
}
\arguments{
\item{x}{regularly sampled univariate time series.  Can be passed as a ts 
object, a data.frame or simple vector of values.  In the case of the latter, deltat must be specified.
If passed as a data.frame, the first column must provide timing, the second the data itself.}

\item{y}{regularly sampled univariate time series.  Can be passed as a ts 
object, a data.frame or simple vector of values.  In the case of the latter, deltat must be specified.
If passed as a data.frame, the first column must provide timing, the second the data itself.  Must share sample times with x.}

\item{deltat}{time interval between samples, used only if x is passed as a vector.  Defaults to 1.}

\item{nw}{multitaper frequency bandwidth parameter, defaults to 4.}

\item{k}{number of tapers, typically 2 * nw - 1, defaults to 7.}

\item{nPoly}{order of the polynomial used for DC/trend removal, defaults to 3.}
}
\value{
list containing frequency, bispectrum, bispectral power, and bicoherence.
 \item{"freq1"}{f1 frequencies (0 to Nyquist)}
 \item{"freq2"}{f12 frequencies (-Nyquist to Nyquist / 2)}
 \item{"bispec"}{Complex bispectrum [f1, f2]}
 \item{"coher"}{bicoherence [f1, f2]}
 \item{"bpspec"}{bispectral power [f1,f2]}
}
\description{
Compute cross-bispectrum with the multitaper method
}
\examples{
t <- 1:256
f <- c( 1/20, 1/5 )
s <- sin( 2 * pi * t \%o\% f )
x <- apply( s, 1, sum ) + rnorm( length( t ), sd = 0.25 )
y <- sin( 2 * pi * t / 4 ) + rnorm( length( t ), sd = 0.25 )
b <- xbispec.mtm( x = x, y = y, deltat = 1 )
image( b$freq1, b$freq2, t( b$coher ), useRaster = TRUE )
}
