% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttools.R
\name{zpad}
\alias{zpad}
\title{Zero pad}
\usage{
zpad(s, n)
}
\arguments{
\item{s}{vector of time series samples}

\item{n}{expansion factor - returned vector will be at least n times original length.}
}
\value{
Padded input signal
}
\description{
Pad a vector with zeros
}
\examples{
n <- 16
s <- rep( 1, 32 )
spad <- zpad( s, n )
S <- fft( spad )
# Show 
f <- seq( 0, to = 0.5, length.out = length( spad ) )
plot( f, Mod( S[1:(length( spad ) / 2 + 1 )] ), type = "l", xlab = "Frequency", ylab = "Mod(S)" )

}
